library update

+++ modules/DeepLearn/RS2.py
@@ -28,7 +28,7 @@ class RS2_predict():
                 self.out_list.append("{}/{}{}".format(output_folder, ln_txt[10:], "_0000.nii.gz"))
                 # self.out_list.append(ln_txt[10:] + "_0000.nii.gz")
 
-    def output_files(self: 'list_path'):
+    def output_files(self) -> list[None]:
         return self.out_list
 
 ##############################################################################
@@ -53,17 +53,19 @@ class RS2_predict_datamanagement():
         if not os.path.exists(tmp_folder):
             os.makedirs(tmp_folder)
 
+        if not isinstance(list_files, list):
+            list_files = [list_files]
+
         for cur_file in list_files:
             if cur_file.endswith('.nii'):
-                with open(cur_file, 'rb') as f_in:
-                    print('Processing in progress for {}'.format(cur_file))
-                    tmp = os.path.basename(cur_file)
-                    lsfield = tmp.split('.')
-                    name = ('.').join(tmp.split('.')[:-1])
-                    ext = tmp.split('.')[-1]
-                    new_name_file = os.path.join(tmp_folder, name + '_0000.nii')
-                    with gzip.open(new_name_file + '.gz', 'wb') as f_out:
-                        shutil.copyfileobj(f_in, f_out)
+                if os.path.exists(cur_file):
+                    with open(cur_file, 'rb') as f_in:
+                        print('Processing in progress for {}'.format(cur_file))
+                        tmp = os.path.basename(cur_file)
+                        name = ('.').join(tmp.split('.')[:-1])
+                        new_name_file = os.path.join(tmp_folder, name + '_0000.nii')
+                        with gzip.open(new_name_file + '.gz', 'wb') as f_out:
+                            shutil.copyfileobj(f_in, f_out)
 
         lso = ["RS2_predict", "-i", tmp_folder, '-o', output_folder, '-m', pretrained_model_path]
         for ef, ev in options.items():
@@ -85,5 +87,5 @@ class RS2_predict_datamanagement():
                 # self.out_list.append(ln_txt[10:] + "_0000.nii.gz")
         shutil.rmtree(tmp_folder)
 
-    def output_files(self: 'list_path'):
+    def output_files(self) -> list[None]:
         return self.out_list
